Make it build on NetBSD 7.0.2, cast *char to int
gcc 4.8.4 on NetBSD 7.0.2 gives a warning and wants the argument to
isspace(), isdigit() and tolower() be an int, so cast the *char values.
-Werror is used, so it did the right thing and aborted.

4721927c-acd0-11e6-af0d-6711e8bb1b6b

+++ src/string.c
@@ -87,7 +87,7 @@ char *string_to_lower(char *str)
 		return p;
 
 	while (*p) {
-		*p = tolower(*p);
+		*p = tolower((int)*p);
 		p++;
 	}
 
@@ -106,7 +106,7 @@ char *trim_str_front(char *dest)
 	assert(dest);
 
 	size = strlen(dest);
-	while (p < dest + size && isspace(*p))
+	while (p < dest + size && isspace((int)*p))
 		p++;
 	if (p == dest)
 		return dest; /* Nothing needs to be done */
@@ -131,7 +131,7 @@ char *trim_str_end(char *dest)
 	if (!size)
 		return dest;
 	p = dest + size - 1;
-	while (p > dest && isspace(*p))
+	while (p > dest && isspace((int)*p))
 		*p-- = '\0';
 
 	return dest;
